move forEachDocumentField as a reusable schema helper
